Update Gatsby and dependencies versions
This commit updates the Gatsby version to 5.12.12 and the gatsby-source-contentful version to 8.12.5 in the package.json file. It also updates the prettier devDependency version to 3.1.1. These updates ensure compatibility with the latest features and bug fixes in Gatsby and its associated plugins, and improve code formatting with the updated prettier version. No significant issues or considerations arise from these updates.